WorkingCodeWithoutDeleteBook

+++ src/main/java/com/barclays/service/BookServiceImpl.java
@@ -1,9 +1,12 @@
 package com.barclays.service;
 
 import com.barclays.model.Book;
+import com.barclays.model.BorrowedBook;
 import com.barclays.repository.BookRepository;
+import com.barclays.repository.BorrowedBookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,8 +18,12 @@ public class BookServiceImpl implements BookService {
     @Autowired
     private final BookRepository bookRepository;
 
-    public BookServiceImpl(BookRepository bookRepository) {
+    @Autowired
+    private final BorrowedBookRepository borrowedBookRepository;
+
+    public BookServiceImpl(BookRepository bookRepository, BorrowedBookRepository borrowedBookRepository) {
         this.bookRepository = bookRepository;
+        this.borrowedBookRepository = borrowedBookRepository;
     }
 
     @Override
@@ -42,7 +49,20 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
+    @Transactional
     public void deleteBook(Book book) {
+        if (book == null || book.getId() == null) {
+            throw new IllegalArgumentException("Book or Book ID must not be null");
+        }
+
+        Integer bookId = book.getId();
+
+        List<BorrowedBook> borrowedBooks = borrowedBookRepository.findByBookId(bookId);
+
+        if (!borrowedBooks.isEmpty()) {
+            throw new IllegalStateException("Cannot delete book as it has borrowed records.");
+        }
+
         bookRepository.delete(book);
     }
